migrate shortURL

+++ db/record/record.go
@@ -10,7 +10,7 @@ import (
 )
 
 // Add adds a record to the database.
-func Add(ctx context.Context, code string, target string, userid int) (*ent.Record, error) {
+func Add(ctx context.Context, code string, target string, userid int64) (*ent.Record, error) {
 	user, err := db.Client.User.Query().
 		Where(user.Tgid(userid)).
 		Only(ctx)
