+++ ent/schema/user.go
@@ -15,7 +15,7 @@ type User struct {
 func (User) Fields() []ent.Field {
 	return []ent.Field{
 		field.String("username").NotEmpty(),
-		field.Int("tgid").Unique().Immutable(),
+		field.Int64("tgid").Unique().Immutable(),
 
 		// permissions
 		field.Bool("customCode").Default(false).Comment("If the user can use custom code"),
